Rework of processing the backend data
Regarding #82
Created new file to only do the "calculations", rather it extracts the data from the huge arrays we are recieving.
Thus removed all the calculations from the BackendConnect.
! Implemented callback funtion, which calls the setState in the DetailsComponent. With that we can display the LoadingComponent, while waiting for the backend data.
Once the postCalculationRequest is finished, we call the callback funtion defined in DetailsComponent, which sets the stillLoading variable to false.
React components automatically retrigger the render function if the state is changed.
Removed all the stillLoading stuff in the charts and tables.